Revert "Feature/issue 125 cholesky derivative"
